FCT-PostScore
Ajout de post score pour envoyer un score à la fin d'une partie au serveur et qu'il soit ajouté à la BD

+++ ServerFlappybirb/Models/Score.cs
@@ -8,7 +8,7 @@
 
         public string? date { get; set; }
 
-        public int timeInSeconds { get; set; }
+        public double timeInSeconds { get; set; }
 
         public int scoreValue { get; set; }
